backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
authlib==1.3.0
httpx==0.26.0
qrcode[pil]==7.4.2
websockets==12.0
python-dotenv==1.0.0
itsdangerous==2.1.2
